requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
lxml==5.3.2
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2